requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
psutil==5.9.7
python-bitcoinlib==0.12.2
pyzmq==25.1.2
requests==2.31.0
urllib3==2.1.0
